requirements.txt
Django==1.6.2
factory-boy==2.3.1
